+++ src/domain/access/role-set/types/agent.role.key.ts
@@ -1,0 +1,7 @@
+import { AgentInfo } from '@core/authentication.agent.info/agent.info';
+import { IRoleSet } from '../role.set.interface';
+
+export type AgentRoleKey = {
+  agentInfo: AgentInfo;
+  roleSet: IRoleSet;
+};
