+++ src/domain/access/role-set/types/roleset.community.membership.status.ts
@@ -1,0 +1,6 @@
+import { CommunityMembershipStatus } from '@common/enums/community.membership.status';
+
+export type RoleSetCommunityMembershipStatus = {
+  roleSetId: string;
+  membershipStatus: CommunityMembershipStatus;
+};
